// packages/trading-core/src/price-service/enigma/types/generated/enigma-price-service.ts
export interface PriceData {
  name: string;
  markPrice: number;
  time: number;
}